[MOBILE] add feature_goods_ride_booking, feature_goods_ride, unit test usecase also viewmodel and fix some bug

+++ mobile_kotlin/nebeng/app/src/main/java/com/example/nebeng/feature_passenger_ride_booking/data/remote/model/dto/DriverDto.kt
@@ -25,7 +25,7 @@ data class DriverDto(
     @field:SerializedName("driver_license_rejected_reason") val driverLicenseRejectedReason: String,
     @field:SerializedName("id_card_fullname") val idCardFullname: String,
     @field:SerializedName("telephone") val telephone: String,
-    @field:SerializedName("average_rating") val averageRating: Double,
+    @field:SerializedName("average_rating") val averageRating: Float,
     @field:SerializedName("id_card_verified") val idCardVerified: Boolean,
     @field:SerializedName("rating_count") val ratingCount: Int,
     @field:SerializedName("profile_img") val profileImg: String?,
